Rename basic tests to `basic.js`
