+++ test/all.js
@@ -1,2 +1,2 @@
-require('.')
+require('./basic')
 require('./web')
